corosync: Do not regenerate authkey if file exists
To decide whether to generate the authkey, we simply check if the
attribute is set on the node. If it's not but the authkey file exists,
we should simply re-use the existing authkey and import it to the
attribute.

This will allow changing the founder node to another node of the
cluster.