Fix race on initial run where cluster founder is not found
On the very first time we apply a pacemaker proposal, non-founder nodes
might be running chef-client faster than the founder node (depending on
how crowbar will order them) and this might result in the founder
attribute not being indexed fast enough. So the nodes will fail not
finding the founder.

So on this very first time, we wait a bit to make sure the founder will
be returned in our search results.