crowbar: Workaround to avoid the drift caused by sync marks (bsc#935462)
Because each chef-client runs executes everything in all cookbooks, and because
the crm calls slow things down considerably, we have a drift issue where the
founder (which does much more work since it's the only one doing the crm calls)
goes much slower than the other nodes. This can impact things badly when it's
so slow that it triggers the timeout in the sync marks.

The workaround here is to reset the sync marks on the founder, so that all
nodes wait on the sync marks and the drift is only happening between sync
marks.

The goal is to only do this when applying a proposal, so that other chef-client
runs are not blocked waiting for sync marks.

Unfortunately, there are cases where we cannot do that: for swift, it's
likely that swift-ring-compute would be on a cluster node, and since
that role is run first, the cluster node would be alone in running chef
and blocking forever. This means that this small helper here will have
to be called explicitly by all barclamps for which there's no such issue
-- afaict, swift is the only one with an issue.

All of this should be removed when we get proper crowbar orchestration.

https://bugzilla.suse.com/show_bug.cgi?id=935462